move identity search to its own file
